requirements.txt
requests
torch
torchvision
Pillow
numpy
scipy
skimage
opencv-python
